imagenes para categorias

+++ app/Http/Controllers/Product/CategoryController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Image as Img;
+use File;
 
 class CategoryController extends Controller
 {
@@ -40,14 +42,38 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-
         $this->validate($request,Category::$rules,Category::$messages);
 
-        //registrar nueva categoria
-        $category = Category::create($request->all());
+        $category = Category::create($request->only('name','description'));
 
-        $notificacion = 'La categoría fue creada correctamente.';
-        $status = 'success';
+        if($request->hasFile('image')){
+            $image = $request->file('image');
+            $filename = uniqid() . $image->getClientOriginalName();
+            $path = public_path("images\categories\\" . $filename);
+
+            $moved = Img::make($image)
+                ->resize(250,250)
+                ->save($path);
+
+            if($moved){
+                $category->image = 'images/categories/' . $filename;
+
+                $category->save();
+
+                $notificacion = 'La categoría fue creada correctamente.';
+                $status = 'success';
+            }else{
+                $notificacion = 'Hubo un problema al guardar la imágen';
+                $status = 'error';
+                return redirect()->route('categories.index')->with(compact('notificacion','status'));
+            }
+
+        }else{
+
+            $notificacion = 'La categoría fue creada correctamente.';
+            $status = 'success';
+        }
+
 
         return redirect()->route('categories.index')->with(compact('notificacion','status'));
     }
@@ -92,10 +118,47 @@ class CategoryController extends Controller
         $this->validate($request,Category::$rules,Category::$messages);
 
         //registrar nueva categoria
-        $category->update($request->all());
+        $category->update($request->only('name','description'));
 
-        $notificacion = 'La categoría fue editada correctamente.';
-        $status = 'success';
+        if($request->hasFile('image')){
+            if($category->image){
+                $deleted = true;
+                $fullPath = public_path().'\\'.$category->image;
+                $deleted = File::delete($fullPath);
+
+                if(!$deleted){
+                    $notificacion = 'Hubo un problema al eliminar la imagen anterior';
+                    $status = 'error';
+                    return redirect()->route('categories.show',$category->id)->with(compact('notificacion','status'));
+                }
+            }
+
+            $image = $request->file('image');
+            $filename = uniqid() . $image->getClientOriginalName();
+            $path = public_path("images\categories\\" . $filename);
+
+            $moved = Img::make($image)
+                ->resize(250,250)
+                ->save($path);
+
+            if($moved){
+                $category->image = 'images/categories/' . $filename;
+
+                $category->save();
+
+                $notificacion = 'La categoría fue editada correctamente.';
+                $status = 'success';
+            }else{
+                $notificacion = 'Hubo un problema al guardar la imágen';
+                $status = 'error';
+                return redirect()->route('categories.index')->with(compact('notificacion','status'));
+            }
+
+        }else{
+
+            $notificacion = 'La categoría fue editada correctamente.';
+            $status = 'success';
+        }
 
         return redirect()->route('categories.index')->with(compact('notificacion','status'));
     }
@@ -108,7 +171,19 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+
+        $deleted = true;
+        $fullPath = public_path().'/'.$category->image;
+        $deleted = File::delete($fullPath);
+
+        if(!$deleted){
+            $notificacion = 'La categoría no pudo ser eliminada.';
+            $status = 'error';
+            return redirect()->back()->with(compact('notificacion','status'));
+        }
+
         $category->delete();
+
         $notificacion = 'La categoría fue eliminada correctamente.';
         $status = 'success';
         return redirect()->back()->with(compact('notificacion','status'));
